Refactor: divide PostJobLimitDto to req, res
