+++ src/main/java/com/spring/tming/domain/post/dto/request/PostUpdateReq.java
@@ -1,6 +1,5 @@
 package com.spring.tming.domain.post.dto.request;
 
-import com.spring.tming.domain.post.dto.PostJobLimitDto;
 import com.spring.tming.global.meta.Skill;
 import java.sql.Timestamp;
 import java.util.List;
@@ -16,7 +15,7 @@ public class PostUpdateReq {
     private String title;
     private String content;
     private Timestamp deadline;
-    private List<PostJobLimitDto> jobLimits;
+    private List<PostJobLimitReq> jobLimits;
     private List<Skill> skills;
 
     @Builder
@@ -25,7 +24,7 @@ public class PostUpdateReq {
             String title,
             String content,
             Timestamp deadline,
-            List<PostJobLimitDto> jobLimits,
+            List<PostJobLimitReq> jobLimits,
             List<Skill> skills) {
         this.postId = postId;
         this.title = title;
