+++ src/main/java/com/spring/tming/domain/post/dto/response/PostJobLimitRes.java
@@ -1,4 +1,4 @@
-package com.spring.tming.domain.post.dto;
+package com.spring.tming.domain.post.dto.response;
 
 import com.spring.tming.global.meta.Job;
 import lombok.AccessLevel;
@@ -8,12 +8,12 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostJobLimitDto {
+public class PostJobLimitRes {
     private Job job;
     private Integer headcount;
 
     @Builder
-    private PostJobLimitDto(Job job, Integer headcount) {
+    private PostJobLimitRes(Job job, Integer headcount) {
         this.job = job;
         this.headcount = headcount;
     }
